Update/exceptions and tests

+++ project/map-backend/src/main/java/com/stevenst/app/repository/MarkerRepository.java
@@ -3,6 +3,6 @@ package com.stevenst.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.stevenst.app.model.Marker;
 
-public interface MarkerRepository extends JpaRepository<Marker, Integer> {
+public interface MarkerRepository extends JpaRepository<Marker, Long> {
 	
 }
